requirements.txt
# Multi-Agent System für automatische LinkedIn-Posts
# XRechnung mit invory.de Integration

# Multi-Agent Framework
crewai==0.1.0
langchain==0.1.0
langchain-openai==0.0.2

# LinkedIn API
linkedin-api==2.0.0

# HTTP Requests
requests==2.31.0
aiohttp==3.9.1
beautifulsoup4==4.12.2
lxml==4.9.3

# Scheduling
schedule==1.2.0
python-dotenv==1.0.0

# Data Processing
pandas==2.1.3

# LLM Providers
openai==1.6.1

# Utilities
python-dateutil==2.8.2
pydantic==2.5.2

